+++ src/index.ts
@@ -7,10 +7,7 @@ export {
   WaitFirstOptions,
   WaitAllOptions,
   SetWorkflowDelayOptions,
-  PreparedWorkflow,
 } from './dbos';
-
-export { PrepareEnqueuedWorkflowOptions } from './dbos-executor';
 
 export { DBOSClient } from './client';
